requirements.txt
Flask~=2.2.2
requests~=2.28.2
requests-oauthlib==1.0.0
Authlib~=1.2.0
gunicorn==20.1.0
python-dotenv==0.21.1
pymongo~=4.3.3